+++ pysgrs/alphabets/__init__.py
@@ -1,1 +1,2 @@
 from pysgrs.alphabets.base import *
+from pysgrs.alphabets.specific import *